PRSD-670: Move the invitation url builder to AbsoluteUrlProvider
