+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/services/AbsoluteUrlProvider.kt
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service
 import org.springframework.web.servlet.mvc.method.annotation.MvcUriComponentsBuilder
 import org.springframework.web.servlet.mvc.method.annotation.MvcUriComponentsBuilder.on
 import uk.gov.communities.prsdb.webapp.controllers.LandlordDashboardController
+import uk.gov.communities.prsdb.webapp.controllers.RegisterLAUserController
 import java.net.URI
 
 @Service
@@ -11,6 +12,12 @@ class AbsoluteUrlProvider {
     fun buildLandlordDashboardUri(): URI =
         MvcUriComponentsBuilder
             .fromMethodCall(on(LandlordDashboardController::class.java).index())
+            .build()
+            .toUri()
+
+    fun buildInvitationUri(token: String): URI =
+        MvcUriComponentsBuilder
+            .fromMethodCall(on(RegisterLAUserController::class.java).acceptInvitation(token))
             .build()
             .toUri()
 }
